refactor(tests): split samples.test.ts into ihex/srec, rename parser test to ihex-parser, fix SAMPLE.md duplicate
